Don't datafy namespaces and classes
if we do, since we no longer elide the first level of the datafied 
representation this results in an endless loop for unreasons still 
unclear. The endless loop is triggered a.o. when unrepl prints out the 
current namespace (*ns*) within :prompt messages.

Printing classes with datafy takes an unresonable amount of time under 
some circumstances (e.g. when nextjournal socket server (on 5.5.5.5) is 
upgraded with unrepl. This don't happen within regular (Docker) 
runtimes.